+++ CHANGELOG.md
@@ -16,10 +16,10 @@ and simply didn't have the time to go back and retroactively create one.
 - Changed zsh prompt to match CWD of other shell prompts
 ### Added
 - Added `ssl-bind` and `ssl-connect` channel protocols for encrypted shells
-- Added `--certificate/--cert` argument to entrypoint and `connect` command
+- Added `ncat`-style ssl arguments to entrypoint and `connect` command
 - Added query-string arguments to connection strings for both the entrypoint
   and the `connect` command.
-  
+
 ## [0.4.2] - 2021-06-15
 Quick patch release due to corrected bug in `ChannelFile` which caused command
 output to be empty in some situations.